2nd Level Expansion
Extended the hallways of the 2nd floor further. Finished the double door hallway with doors on the walls and spotlights going through them. Added a placeholder body in the hallway. Made the ceiling tiles two sided materials and deleted the greybox above the world. Added volumetric fog and centralized it to the level.